Add files via upload
Write a program to print the reverse of the given string.

Input

The input will be a single line containing a string.

Output

The output should be a single line containing the reverse of the given string.

Explanation

For example, if the given string is "Hurray! We have won the match.", your code should print the reverse of the string, which is ".hctam eht now evah ew lyarruH".